Rename methods
