+++ packages/apps/examples/sdk-api/src/controllers/nft.controller.ts
@@ -63,31 +63,4 @@ export default {
       return res.status(500).send({});
     }
   },
-
-  getCollectionEvents: async (req: Request, res: Response) => {
-    try {
-      const collectionEvents = await nfts.getCollectionEvents({
-        contractAddress: req.params.contractAddress,
-        first: 5,
-      });
-      return res.status(200).send(collectionEvents);
-    } catch (error) {
-      console.error(error);
-      return res.status(500).send({});
-    }
-  },
-
-  getNFTEvents: async (req: Request, res: Response) => {
-    try {
-      const NFTEvents = await nfts.getNFTEvents({
-        contractAddress: req.params.contractAddress,
-        tokenId: req.params.tokenId,
-        first: 5,
-      });
-      return res.status(200).send(NFTEvents);
-    } catch (error) {
-      console.error(error);
-      return res.status(500).send({});
-    }
-  },
 };
