+++ packages/libs/sdk/spec/recordings/query-getEventsByContract-base_2069415806/recording.har
@@ -62,18 +62,18 @@
           "url": "https://api.quicknode.com/graphql"
         },
         "response": {
-          "bodySize": 779,
+          "bodySize": 775,
           "content": {
             "encoding": "base64",
             "mimeType": "application/json",
-            "size": 779,
-            "text": "[\"H4sIAAAAAAAA/9WTS0/bQBDHv8ue02p29u0bQtByaMQj4kBVoVnvmASwjWyHBqF8965d0iIoQpV6qeXDeN7z/8mPItFAongUPCy543U92mXbDB2Vw2gP7Q03B/fcDP2Ulq44G18fRdMmHj3xti1v5us6cicKqaT3zsJMVF1b76XUcZ/TBWwkVCTLHOWSgtFaBvJGSqyCM9KYyklHyTOJmRhWNfcD1Xe5EAHlBwj5XYAsTCi0+ggAF2Na+7w/KIxOg7HBSK2U0g4COEvGaYOhSoAoJSkeCztq+nzeqm0+U7+cytEnDh5Re5AYiEMIzBh9sg==\",\"wVuwyjjWKWpJNs+IpXMcefz0pfXKIO7aVtwdNYk3okCfVRge7rJIYnG6Nz87PDidts56HqUsFVo1+yX181NQgi0BqHSJgJRWQGyt8y5IFdgrVOCIiXbdTtbUDKvhIfecicvLcWZD9TR3DC+e9pogiu2LlP2nBRa/OWdXw5M8B5l2rngbtrFGmX8PWy8AC3SFhj/DrvKTqUbNKZGmIGUywagSlbcRK2t9CMqCfAO2A5lxYqWYMm4ZTEg2WnbRKoVZ2+CwtJVEpyQYiMTsnHWQ8i0JXHSvYUsj34Nt/g/Y32bijq74qKna6X9v0v6669tuHIAXdYnncIJhHevz6/Tp9j5et5u82pL6OW+G41wpiqFb8+Q67vh+1a77n+6Kbvvsz7C74b2e38WLzY93O/3VSa+y54eLXcGr2MH5l7NyyTUtRo7b7fYH/OSsDB0FAAA=\"]"
+            "size": 775,
+            "text": "[\"H4sIAAAAAAAA/9WTS0/bQBDHv8ue02p29u0bQtByaMQj4kBVoVnvmASwjWyHBqF8965d0iIoQpV6qeXDeN7z/8mPItFAongUPCy543U92mXbDB2Vw2gP7Q03B/fcDP2Ulq44G18fRdMmHj3xti1v5us6cicKqaT3zsJMVF1b76XUcZ/TBWwkVCTLHOWSgtFaBvJGSqyCM9KYyklHyTOJmRhWNfcD1Xe5EAHlBwj5XYAsTCi0+ggAF2Na+7w/KIxOg7HBSK2U0g4COEvGaYOhSoAoJSkeCztq+nzeqm0+U7+cytEnDh5Re5AYiEMI\",\"zBh9ssFbsMo41ilqSTbPiKVzHHn89KX1yiDu2lbcHTWJN6JAn1UYHu6ySGJxujc/Ozw4nbbOeh6lLBVaNfsl9fNTUIItAah0iYCUVkBsrfMuSBXYK1TgiIl23U7W1Ayr4SH3nInLy3FmQ/U0dwwvnvaaIIrti5T9pwUWvzlnV8OTPAeZdq54G7axRpl/D1svAAt0hYY/w67yk6lGzSmRpiBlMsGoEpW3EStrfQjKgnwDtgOZcWKlmDJuGUxINlp20SqFWdvgsLSVRKckGIjE7Jx1kPItCVx0r2FLI9+Dbf4P2N9m4o6u+Kip2ul/b9L+uuvbbhyAF3WJ53CCYR3r8+v06fY+XrebvNqS+jlvhuNcKYqhW/PkOu74ftWu+5/uim777M+wu+G9nt/Fi82Pdzv91UmvsueHi13Bq9jB+Zezcsk1LUaO2+32B/zkrAwdBQAA\"]"
           },
           "cookies": [],
           "headers": [
             {
               "name": "date",
-              "value": "Wed, 31 May 2023 19:19:04 GMT"
+              "value": "Thu, 01 Jun 2023 14:11:42 GMT"
             },
             {
               "name": "content-type",
@@ -117,7 +117,7 @@
             },
             {
               "name": "x-ratelimit-rpmreset",
-              "value": "1685560800"
+              "value": "1685628760"
             },
             {
               "name": "x-ratelimit-rpslimit",
@@ -129,7 +129,7 @@
             },
             {
               "name": "x-ratelimit-rpsreset",
-              "value": "1685560744"
+              "value": "1685628702"
             },
             {
               "name": "cf-cache-status",
@@ -137,7 +137,7 @@
             },
             {
               "name": "report-to",
-              "value": "{\"endpoints\":[{\"url\":\"https:\\/\\/a.nel.cloudflare.com\\/report\\/v3?s=9yBCNJWHinWk0vq7BrgwDG7kFuT4OCTbvW0VbzKmLRwXmTW%2Bfiu8hNUE2%2BUXgleHwg1UqMQlTduqFDxHXAh%2FS%2FerPOGfNdwZ5HLlwuger3P5U5N9wM3P2YxzwX4xaBIW9wRf\"}],\"group\":\"cf-nel\",\"max_age\":604800}"
+              "value": "{\"endpoints\":[{\"url\":\"https:\\/\\/a.nel.cloudflare.com\\/report\\/v3?s=ZsjnLNJQCvbMqyiMcJiBSfpSDzVG9ye%2BKhHZE6QJt3lcm9bC%2F12dpQdu%2B%2FumTMnD5EbwOmWvp0MFY8ZKl7bC7hg5YMW%2FqWIM9ucEupJAAiivLHt1ZUsyt2QxTzCsr7bL%2BcUm\"}],\"group\":\"cf-nel\",\"max_age\":604800}"
             },
             {
               "name": "nel",
@@ -165,17 +165,17 @@
             },
             {
               "name": "cf-ray",
-              "value": "7d0190f83a78429a-EWR"
+              "value": "7d080c18888e42a3-EWR"
             }
           ],
-          "headersSize": 1106,
+          "headersSize": 1110,
           "httpVersion": "HTTP/1.1",
           "redirectURL": "",
           "status": 200,
           "statusText": "OK"
         },
-        "startedDateTime": "2023-05-31T19:19:03.512Z",
-        "time": 756,
+        "startedDateTime": "2023-06-01T14:11:41.263Z",
+        "time": 1224,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -183,7 +183,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 756
+          "wait": 1224
         }
       }
     ],
